backend/pyproject.toml
[project]
name = "408-robot-backend"
version = "0.1.0"
description = "408 考研辅助 Agent 后端服务"
requires-python = ">=3.12"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    # Vector DB
    "chromadb>=0.5.0",
    # Embeddings
    "sentence-transformers>=3.0.0",
    # LangChain (v0.3+)
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-chroma>=0.2.0",
    # Document generation
    "python-docx>=1.1.0",
    # Database
    "aiosqlite>=0.20.0",
    # Utilities
    "httpx>=0.27.0",
    "tqdm>=4.60.0",
    "python-multipart>=0.0.12",
]

[project.scripts]
start = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"

[tool.uv]
python-downloads = "automatic"
